Have added manager indicator for "your departments", along with "none assigned" for Your departments if emtpy. Also added permission checking in the ajax files. Changed the name of the editmode buttons to better suit. Cleaned up a tiny bit of the ordering and any missing maxlength limits.

+++ ajax/deleteuser.ajax.php
@@ -2,7 +2,8 @@
 session_start();
 require_once '../includes/functions.inc.php';
 require_once '../includes/dbh.inc.php';
-if (isset($_POST["UserID"])) {
+
+if (isset($_POST["UserID"])  && $_SESSION["role"] == "3") {//make sure the request has the correct info, and the user has the correct perms
     if (!deleteUser($conn, $_POST, $_SESSION)) {
         echo json_encode(array("status" => "fail - could not delete user"));
         return;
